Reactive Hibernate now uses the latest Postgres version
