--- sql-db/hibernate-reactive/src/test/java/io/quarkus/ts/reactive/Postgresql13DatabaseIT.java
@@ -1,33 +1,0 @@
-package io.quarkus.ts.reactive;
-
-import io.quarkus.test.bootstrap.PostgresqlService;
-import io.quarkus.test.bootstrap.RestService;
-import io.quarkus.test.scenarios.QuarkusScenario;
-import io.quarkus.test.services.Container;
-import io.quarkus.test.services.QuarkusApplication;
-
-@QuarkusScenario
-public class Postgresql13DatabaseIT extends AbstractReactiveDatabaseIT {
-
-    private static final String POSTGRES_USER = "quarkus_test";
-    private static final String POSTGRES_PASSWORD = "quarkus_test";
-    private static final String POSTGRES_DATABASE = "quarkus_test";
-    static final int POSTGRES_PORT = 5432;
-
-    @Container(image = "${postgresql.13.image}", port = POSTGRES_PORT, expectedLog = "listening on IPv4 address")
-    static PostgresqlService database = new PostgresqlService()
-            .withUser(POSTGRES_USER)
-            .withPassword(POSTGRES_PASSWORD)
-            .withDatabase(POSTGRES_DATABASE);
-
-    @QuarkusApplication
-    static RestService app = new RestService().withProperties("postgresql.properties")
-            .withProperty("quarkus.datasource.username", POSTGRES_USER)
-            .withProperty("quarkus.datasource.password", POSTGRES_PASSWORD)
-            .withProperty("quarkus.datasource.reactive.url", database::getReactiveUrl);
-
-    @Override
-    protected RestService getApp() {
-        return app;
-    }
-}
