+++ sql-db/hibernate-reactive/src/test/java/io/quarkus/ts/reactive/PostgresqlDatabaseIT.java
@@ -1,26 +1,20 @@
-package io.quarkus.ts.reactive.openshift;
-
-import org.junit.jupiter.api.Disabled;
-import org.junit.jupiter.api.condition.EnabledIfSystemProperty;
+package io.quarkus.ts.reactive;
 
 import io.quarkus.test.bootstrap.PostgresqlService;
 import io.quarkus.test.bootstrap.RestService;
-import io.quarkus.test.scenarios.OpenShiftScenario;
+import io.quarkus.test.scenarios.QuarkusScenario;
 import io.quarkus.test.services.Container;
 import io.quarkus.test.services.QuarkusApplication;
-import io.quarkus.ts.reactive.AbstractReactiveDatabaseIT;
 
-@OpenShiftScenario
-@EnabledIfSystemProperty(named = "ts.redhat.registry.enabled", matches = "true")
-//TODO: enable disabled methods then the issue will be fixed
-public class OpenShiftPostgresql12IT extends AbstractReactiveDatabaseIT {
+@QuarkusScenario
+public class PostgresqlDatabaseIT extends AbstractReactiveDatabaseIT {
 
     private static final String POSTGRES_USER = "quarkus_test";
     private static final String POSTGRES_PASSWORD = "quarkus_test";
     private static final String POSTGRES_DATABASE = "quarkus_test";
     private static final int POSTGRES_PORT = 5432;
 
-    @Container(image = "${postgresql.12.image}", port = POSTGRES_PORT, expectedLog = "listening on IPv4 address")
+    @Container(image = "${postgresql.13.image}", port = POSTGRES_PORT, expectedLog = "listening on IPv4 address")
     static PostgresqlService database = new PostgresqlService()
             .withUser(POSTGRES_USER)
             .withPassword(POSTGRES_PASSWORD)
@@ -35,15 +29,5 @@ public class OpenShiftPostgresql12IT extends AbstractReactiveDatabaseIT {
     @Override
     protected RestService getApp() {
         return app;
-    }
-
-    @Override
-    @Disabled("https://github.com/quarkusio/quarkus/issues/21700")
-    public void connectToUniEndpoint() {
-    }
-
-    @Override
-    @Disabled("https://github.com/quarkusio/quarkus/issues/21700")
-    public void connectToMultiEndpoint() {
     }
 }
